move files into explore and remove unused files
