+++ pages/explore/TemplateSelect.js
